gdb: rename hi to miri; add `miri run` as the staring command
$ OSDK_LOCAL_DEV=1 rust-gdb -x miri.gdb --args cargo osdk miri test

[inferior=11 pid=24546] [stop_handler] /root/.rustup/toolchains/nightly-2025-12-06-x86_64-unknown-linux-gnu/bin/miri
CmdLine(cmdline='/root/.rustup/toolchains/nightly-2025-12-06-x86_64-unknown-linux-gnu/bin/miri '
                '--sysroot /root/.cache/miri --crate-name os_osdk_bin '
                '--edition=2024 src/main.rs --diagnostic-width=160 '
                '--crate-type bin --emit=dep-info,link -C embed-bitcode=no -C '
                'debuginfo=2 --check-cfg cfg(docsrs,test) --check-cfg '
                'cfg(feature, values()) -C metadata=3c55ea96a279459d -C '
                'extra-filename=-8dc6884ab917371a --out-dir '
                '/KMiri/tests/os/target/miri/x86_64-unknown-linux-gnu/debug/deps '
                '--target x86_64-unknown-linux-gnu -C '
                'incremental=/KMiri/tests/os/target/miri/x86_64-unknown-linux-gnu/debug/incremental '
                '-L '
                'dependency=/KMiri/tests/os/target/miri/x86_64-unknown-linux-gnu/debug/deps '
                '-L dependency=/KMiri/tests/os/target/miri/debug/deps --extern '
                'noprelude:alloc=/KMiri/tests/os/target/miri/x86_64-unknown-linux-gnu/debug/deps/liballoc-5948990b7399f447.rlib '
                '--extern '
                'noprelude:compiler_builtins=/KMiri/tests/os/target/miri/x86_64-unknown-linux-gnu/debug/deps/libcompiler_builtins-71d74021048da112.rlib '
                '--extern '
                'noprelude:core=/KMiri/tests/os/target/miri/x86_64-unknown-linux-gnu/debug/deps/libcore-c06e9e878cc0a338.rlib '
                '--extern '
                'os=/KMiri/tests/os/target/miri/x86_64-unknown-linux-gnu/debug/deps/libos-7481199661742493.rlib '
                '--extern '
                'osdk_frame_allocator=/KMiri/tests/os/target/miri/x86_64-unknown-linux-gnu/debug/deps/libosdk_frame_allocator-9272120b690dac3f.rlib '
                '--extern '
                'osdk_heap_allocator=/KMiri/tests/os/target/miri/x86_64-unknown-linux-gnu/debug/deps/libosdk_heap_allocator-2b88bcb05529d8ae.rlib '
                '--extern '
                'osdk_test_kernel=/KMiri/tests/os/target/miri/x86_64-unknown-linux-gnu/debug/deps/libosdk_test_kernel-d1fa7507f2bd8d2f.rlib '
                '--extern '
                'ostd=/KMiri/tests/os/target/miri/x86_64-unknown-linux-gnu/debug/deps/libostd-4a82d7d6f9c5165c.rlib '
                '-Z unstable-options --cfg ktest -C link-arg=-Tmiri.ld -C '
                'relocation-model=static -C relro-level=off -C '
                'force-unwind-tables=yes --check-cfg cfg(ktest) -C '
                'no-redzone=y -C target-feature=+ermsb -- ',
        exe='/root/.rustup/toolchains/nightly-2025-12-06-x86_64-unknown-linux-gnu/bin/miri')
😎 Reached miri: None
miri
disconnect       run              set-breakpoints
>>> miri run
Breakpoint 2 at 0x5e41f22792f9: file src/bin/miri.rs, line 440.
Breakpoint 3 at 0x5e41f227ff16: file src/eval.rs, line 288.
>>> i b
Num     Type           Disp Enb Address            What
1       catchpoint     keep y                      exec, program "/root/.rustup/toolchains/nightly-2025-12-06-x86_64-unknown-linux-gnu/bin/miri"
        catchpoint already hit 32 times
2       breakpoint     keep y   0x00005e41f22792f9 in miri::main at src/bin/miri.rs:440 inf 11
3       breakpoint     keep y   0x00005e41f227ff16 in miri::eval::create_ecx at src/eval.rs:288 inf 11
>>> c
Continuing.
[Thread debugging using libthread_db enabled]
Using host libthread_db library "/lib/x86_64-linux-gnu/libthread_db.so.1".

Thread 11.1 "miri" hit Breakpoint 2, miri::main () at src/bin/miri.rs:440
440         let early_dcx = EarlyDiagCtxt::new(ErrorOutputType::default());
>>> c
Continuing.
[New Thread 0x7475637ff6c0 (LWP 24580) (id 11.20)]
[New Thread 0x7475635fe6c0 (LWP 24581) (id 11.21)]
[New Thread 0x747562dfd6c0 (LWP 24582) (id 11.22)]
warning: unstable feature specified for `-Ctarget-feature`: `ermsb`
  |
  = note: this feature is not stably supported; its behavior can change in the future

[Switching to Thread 0x7475635fe6c0 (LWP 24581)]

Thread 11.21 "rustc" hit Breakpoint 3, miri::eval::create_ecx (tcx=..., entry_id=..., entry_type=..., config=0x74756356fa98, genmc_ctx=...) at src/eval.rs:288
288         let layout_cx = LayoutCx::new(tcx, typing_env);
