+++ tests/os/.gdb_history
@@ -1,7 +1,3 @@
-r
-da -enabled 
-da -enabled off
-r
 i inferiors 
 c
 i inferiors 
@@ -254,3 +250,7 @@ hi disconnect
 c
 n
 i inferiors 
+miri run
+i b
+c
+c
